+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 * Unreleased
     * Add ability to set source directory for the Avro compiler's Velocity templates.
+    * Matching of fieldVisibility settings is now case-insensitive.
+    * Removed some excessive debug logging
 
 * 0.5.0
     * Add support for schemas/protocols/IDL in subdirectories of `src/main/avro`, etc. (#11)